buildhck: slightly better test organization
