+++ client.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+
+from buildhck.client import client
+
+if __name__ == '__main__':
+    client.main()
